retoques finales

+++ db/migrate/20200301041914_create_policies.rb
@@ -1,10 +1,10 @@
 class CreatePolicies < ActiveRecord::Migration[5.2]
   def change
     create_table :policies do |t|
-      t.float :prime
+      t.int :prime
       t.date :start_date
       t.date :end_date
-      t.float :comision
+      t.int :comision
       t.string :payment_type, limit: 45
       t.references :product, foreign_key: true
       t.references :client, foreign_key: true
